Change Theme to Bootstrap Cyborg (https://bootswatch.com/cyborg/) Database migrated. Now record store url to image in server. Added Details.cshtml

+++ SpeedwayCenter/SpeedwayCenter/Controllers/SpeedwayController.cs
@@ -1,14 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Data.Entity;
+using System.Diagnostics;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
+using System.Web.UI.WebControls;
 using SpeedwayCenter.Models.Entity_Framework;
 using SpeedwayCenter.Models.Repository;
+using Image = System.Drawing.Image;
 
 namespace SpeedwayCenter.Controllers
 {
@@ -38,11 +41,17 @@ namespace SpeedwayCenter.Controllers
         {
             if (file != null)
             {
-                using (var stream = file.InputStream)
-                {
-                    Image photo = new Bitmap(stream);
-                    rider.Photo = photo;
-                }
+                //    var formatPosition = file.FileName.IndexOf('.');
+                //    var format = file.FileName.Substring(formatPosition);
+                //    var path = HttpContext.Server.MapPath("~/Photos/" + rider.Id + format);
+                var serverPath = $"~/Photos/{rider.Id}.png";
+                var path = HttpContext.Server.MapPath(serverPath);
+                file.SaveAs(path);
+                //using (var image = Image.FromStream(file.InputStream))
+                //{
+                //    image.Save(path, ImageFormat.Png);
+                //}
+                rider.Image = serverPath;
             }
             _repository.Add(rider);
             _repository.Save();
@@ -53,10 +62,28 @@ namespace SpeedwayCenter.Controllers
         public ActionResult Delete(int id)
         {
             var entity = _repository.FindBy(rider => rider.Id == id).FirstOrDefault();
-            _repository.Delete(entity);
-            _repository.Save();
-            var records = _repository.GetAll();
-            return View("Index", records);
+            if (entity == null)
+            {
+                var records = _repository.GetAll(); // TODO the same code in two lines - refactor
+                return View("Index", records);
+            }
+            //else // TODO Uncommnet when error page is completed
+            {
+                if (System.IO.File.Exists(entity.Image))
+                {
+                    System.IO.File.Delete(entity.Image);
+                }
+                _repository.Delete(entity);
+                _repository.Save();
+                var records = _repository.GetAll();
+                return View("Index", records);
+            }
+        }
+
+        public ActionResult Details(int id)
+        {
+            var entity = _repository.FindBy(rider => rider.Id == id).FirstOrDefault();
+            return View(entity);
         }
     }
 }
